refactor: rename to FlashMintAaveDelevered + align with FlashMint conventions
Renames the Arbitrum delever-redemption contract from AaveV3DeleveredRedeemer
to FlashMintAaveDelevered and reshapes its external interface to match the
existing FlashMintLeveraged / FlashMintLeveragedAaveFL conventions so SDK
adapters and downstream tooling can reuse the same patterns.

External interface:
  - redeemExactSetForETH(setToken, setAmount, minAmountOutputToken, swapData[])
  - redeemExactSetForERC20(setToken, setAmount, outputToken, minAmountOutputToken, swapData[])
  - getRedeemExactSet(setToken, setAmount, outputToken, swapData[]) view-style
  - constants.ETH_ADDRESS sentinel matches DEXAdapter convention
  - FlashRedeem event with the same indexed fields as FlashMintLeveraged

State variables:
  - setController, aaveV3Pool, debtIssuanceModule, addresses
  - validSetToken modifier (calls setController.isSet) — matches FlashMintDexV5

Removed:
  - The simple `redeem(setToken, amount, recipient)` from the previous draft
    is replaced by `redeemExactSetForERC20(setToken, amount, headlineUnderlying,
    0, [noopSwap...])` — same outcome, one consistent surface.
  - The `address(0) recipient` sentinel — every other FlashMint sends to
    msg.sender, so we do too. SDK encodes the connected wallet as the from
    address.

Constructor signature: now also takes IController (setController) for the
Set-validity check, matching FlashMintLeveragedAaveFL. Deploy helper updated.

Test plan (passes 17/17 on Arbitrum fork at block 459_500_000):
  - Constructor wires immutables (setController, aaveV3Pool, debtIssuanceModule,
    DEX addresses)
  - aToken interface assumption (3 specs)
  - IPool.withdraw(MaxUint256) semantics
  - 7 product × output-token cases via redeemExactSetForETH / redeemExactSetForERC20:
      AAVE2x → AAVE / WETH / ETH / USDC
      LINK2x → LINK / WETH / ETH
    (covering single-component product, multi-component, multi-hop paths,
    no-op SwapData passthrough, USDT-dust handling, and ETH unwrap)
  - 5 safety reverts: zero amount × 2 functions, insufficient output, swap-data
    length mismatch, non-Set token

This is a force-push to the existing PR branch — the previous integration spec
covering the same flows under the old names is replaced with this one.
